ENH fix astype usage to prevent copying
Removed were possible; using copy=False (from utils.fixes) were needed.

Also some C integer type fixes to gradient boosting.

+++ sklearn/linear_model/randomized_l1.py
@@ -49,7 +49,7 @@ def _resample_model(estimator_func, X, y, scaling=.5, n_resampling=200,
                 verbose=max(0, verbose - 1),
                 **params)
             for _ in range(n_resampling)):
-        scores_ += active_set.astype(np.float)
+        scores_ += active_set
 
     scores_ /= n_resampling
     return scores_
